moved files to kebab-case and updated all imports and readme to match it
