+++ apps/shade/src/index.ts
@@ -81,6 +81,6 @@ export {useSimplePagination} from './hooks/use-simple-pagination';
 export * from '@/lib/utils';
 export {cn, debounce, kebabToPascalCase, formatUrl, formatQueryDate, formatTimestamp, formatNumber, formatDuration, formatPercentage, formatDisplayDate, isValidDomain, getYRange, getYRangeWithMinPadding, getYRangeWithLargePadding, calculateYAxisWidth, getRangeDates, getCountryFlag, sanitizeChartData, formatDisplayDateWithRange, centsToDollars, getRangeForStartDate, formatMemberName, getMemberInitials, stringToHslColor, abbreviateNumber} from '@/lib/utils';
 
-export {default as ShadeApp} from './ShadeApp';
-export type {ShadeAppProps} from './ShadeApp';
-export {useFocusContext} from './providers/ShadeProvider';
+export {default as ShadeApp} from './shade-app';
+export type {ShadeAppProps} from './shade-app';
+export {useFocusContext} from './providers/shade-provider';
