+++ apps/shade/src/shade-app.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import React from 'react';
 // import {FetchKoenigLexical} from './global/form/HtmlEditor';
-import ShadeProvider from './providers/ShadeProvider';
+import ShadeProvider from './providers/shade-provider';
 
 /**
  * The className is used to scope the styles of the app to the app's namespace.
