+++ apps/shade/test/.eslintrc.cjs
@@ -3,5 +3,9 @@ module.exports = {
     plugins: ['ghost'],
     extends: [
         'plugin:ghost/test'
-    ]
+    ],
+    rules: {
+        // Enforce a kebab-case (lowercase with hyphens) for all filenames
+        'ghost/filenames/match-regex': ['error', '^[a-z0-9.-]+$', false]
+    }
 };
